fix bug in local HSE potential
The derivative of the enhancement factor Fx(s) in HSE had an incorrect slope
because a factor of s was missing. As a consequence the term 1/s Fx(s) was
not bounded for s -> 0.

Fix derivative of F(s) and H(s), which were lacking the factor s and s^2 respectively.

Rename FT to int_pot in ExchangeOperator.

git-svn-id: http://qboxcode.org/svn/qb/branches/hse-dev@1406 cba15fb0-1239-40c8-b417-11db7ca47a34