First RSH implementation
RSH is implemented as a separate functional (so far).
New variables alpha_RSH, beta_RSH and mu_RSH are added. There is
still duplicated code with HSEFunctional and PBEFunctional.
The InteractionPotential class is removed and ExchangeOperator is
modified to implement the interaction potential depending on the
parameters alpha_RSH, beta_RSH and mu_RSH.
Functionality needs testing for HF and all hybrid functionals.