Change sign of force on constraints and pair force
The change of sign makes the value of the force consistent with
the definition of the force ( -grad(E) )
A positive value of the constraint force tends to increase the
constraint value.
The change of sign of PairForce makes it such that a positive
PairForce value tends to increase the bond length.
The value of the PairForce needed to compensate for a distance
constraint is opposite to the force acting on the constraint.